tests/wpt/tests/css/css-gaps/tentative/grid/grid-gap-decorations-028.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decoration: :visited column-rule-color style is not leaked via getComputedStyle</title>
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<script src="/resources/testharness.js" type="text/javascript"></script>
<script src="/resources/testharnessreport.js" type="text/javascript"></script>
</head>
<body>
<a class="grid-link">
  <div id="target1"></div>
</a>
<style>
  #target1 {
    display: grid;
    column-rule-color: lime;
  }

  grid:link:visited #target1 {
    column-rule-color: blue;
  }
</style>
<script>
  test(function() {
    const gridContainerStyle = window.getComputedStyle(document.querySelector('#target1'));
    const columnRuleColor = gridContainerStyle.getPropertyValue('column-rule-color');
    assert_equals(columnRuleColor, 'rgb(0, 255, 0)');

  }, "`column-rule-color` should be rgb(0, 255, 0)");

  test(function() {
    const gridContainerStyle = window.getComputedStyle(document.querySelector('#target1'));
    const columnRuleColor = gridContainerStyle.getPropertyValue('column-rule-color', ':visited');
    assert_not_equals(columnRuleColor, 'rgb(0, 0, 255)');

  }, "Visited `column-rule-color` should not be blue but it is");
</script>
</body>
</html>
